Conflicts:
	configure
	configure.ac
	dlls/wined3d/surface.c
	dlls/wined3d/swapchain.c
	dlls/wined3d/swapchain_base.c
	dlls/wined3d/wined3d_private.h
	dlls/winex11.drv/bitblt.c
	include/config.h.in
	programs/winecfg/Bg.rc
	programs/winecfg/Cs.rc
	programs/winecfg/Da.rc
	programs/winecfg/De.rc
	programs/winecfg/En.rc
	programs/winecfg/Es.rc
	programs/winecfg/Fi.rc
	programs/winecfg/Fr.rc
	programs/winecfg/Hu.rc
	programs/winecfg/It.rc
	programs/winecfg/Ja.rc
	programs/winecfg/Ko.rc
	programs/winecfg/Lt.rc
	programs/winecfg/Nl.rc
	programs/winecfg/No.rc
	programs/winecfg/Pl.rc
	programs/winecfg/Pt.rc
	programs/winecfg/Ro.rc
	programs/winecfg/Ru.rc
	programs/winecfg/Si.rc
	programs/winecfg/Sv.rc
	programs/winecfg/Tr.rc
	programs/winecfg/Zh.rc

+++ programs/winetest/main.c
@@ -40,6 +40,9 @@
 /* Don't submit the results if more than SKIP_LIMIT tests have been skipped */
 #define SKIP_LIMIT 10
 
+/* Don't submit the results if more than FAILURES_LIMIT tests have failed */
+#define FAILURES_LIMIT 50
+
 struct wine_test
 {
     char *name;
@@ -61,6 +64,8 @@ static int nr_native_dlls;
 static const char whitespace[] = " \t\r\n";
 static const char testexe[] = "_test.exe";
 static char build_id[64];
+static BOOL is_wow64;
+static int failures;
 
 /* filters for running only specific tests */
 static char *filters[64];
@@ -149,25 +154,34 @@ static int running_under_wine (void)
 
 static int check_mount_mgr(void)
 {
-    if (running_under_wine())
-    {
-        HANDLE handle = CreateFileA( "\\\\.\\MountPointManager", GENERIC_READ,
-                                     FILE_SHARE_READ|FILE_SHARE_WRITE, NULL, OPEN_EXISTING, 0, 0 );
-        if (handle == INVALID_HANDLE_VALUE) return FALSE;
-        CloseHandle( handle );
-    }
+    HANDLE handle = CreateFileA( "\\\\.\\MountPointManager", GENERIC_READ,
+                                 FILE_SHARE_READ|FILE_SHARE_WRITE, NULL, OPEN_EXISTING, 0, 0 );
+    if (handle == INVALID_HANDLE_VALUE) return FALSE;
+    CloseHandle( handle );
     return TRUE;
+}
+
+static int check_wow64_registry(void)
+{
+    char buffer[MAX_PATH];
+    DWORD type, size = MAX_PATH;
+    HKEY hkey;
+    BOOL ret;
+
+    if (!is_wow64) return TRUE;
+    if (RegOpenKeyA( HKEY_LOCAL_MACHINE, "Software\\Microsoft\\Windows\\CurrentVersion", &hkey ))
+        return FALSE;
+    ret = !RegQueryValueExA( hkey, "ProgramFilesDir (x86)", NULL, &type, (BYTE *)buffer, &size );
+    RegCloseKey( hkey );
+    return ret;
 }
 
 static int check_display_driver(void)
 {
-    if (running_under_wine())
-    {
-        HWND hwnd = CreateWindowA( "STATIC", "", WS_OVERLAPPEDWINDOW, CW_USEDEFAULT, 0, CW_USEDEFAULT, 0,
-                                   0, 0, GetModuleHandleA(0), 0 );
-        if (!hwnd) return FALSE;
-        DestroyWindow( hwnd );
-    }
+    HWND hwnd = CreateWindowA( "STATIC", "", WS_OVERLAPPEDWINDOW, CW_USEDEFAULT, 0, CW_USEDEFAULT, 0,
+                               0, 0, GetModuleHandleA(0), 0 );
+    if (!hwnd) return FALSE;
+    DestroyWindow( hwnd );
     return TRUE;
 }
 
@@ -198,6 +212,86 @@ static int running_on_visible_desktop (void)
     return IsWindowVisible(desktop);
 }
 
+static int running_as_admin (void)
+{
+    PSID administrators = NULL;
+    SID_IDENTIFIER_AUTHORITY nt_authority = { SECURITY_NT_AUTHORITY };
+    HANDLE token;
+    DWORD groups_size;
+    PTOKEN_GROUPS groups;
+    DWORD group_index;
+
+    /* Create a well-known SID for the Administrators group. */
+    if (! AllocateAndInitializeSid(&nt_authority, 2, SECURITY_BUILTIN_DOMAIN_RID,
+                                   DOMAIN_ALIAS_RID_ADMINS, 0, 0, 0, 0, 0, 0,
+                                   &administrators))
+        return -1;
+
+    /* Get the process token */
+    if (! OpenProcessToken(GetCurrentProcess(), TOKEN_QUERY, &token))
+    {
+        FreeSid(administrators);
+        return -1;
+    }
+
+    /* Get the group info from the token */
+    groups_size = 0;
+    GetTokenInformation(token, TokenGroups, NULL, 0, &groups_size);
+    groups = heap_alloc(groups_size);
+    if (groups == NULL)
+    {
+        CloseHandle(token);
+        FreeSid(administrators);
+        return -1;
+    }
+    if (! GetTokenInformation(token, TokenGroups, groups, groups_size, &groups_size))
+    {
+        heap_free(groups);
+        CloseHandle(token);
+        FreeSid(administrators);
+        return -1;
+    }
+    CloseHandle(token);
+
+    /* Now check if the token groups include the Administrators group */
+    for (group_index = 0; group_index < groups->GroupCount; group_index++)
+    {
+        if (EqualSid(groups->Groups[group_index].Sid, administrators))
+        {
+            heap_free(groups);
+            FreeSid(administrators);
+            return 1;
+        }
+    }
+
+    /* If we end up here we didn't find the Administrators group */
+    heap_free(groups);
+    FreeSid(administrators);
+    return 0;
+}
+
+static int running_elevated (void)
+{
+    HANDLE token;
+    TOKEN_ELEVATION elevation_info;
+    DWORD size;
+
+    /* Get the process token */
+    if (! OpenProcessToken(GetCurrentProcess(), TOKEN_QUERY, &token))
+        return -1;
+
+    /* Get the elevation info from the token */
+    if (! GetTokenInformation(token, TokenElevation, &elevation_info,
+                              sizeof(TOKEN_ELEVATION), &size))
+    {
+        CloseHandle(token);
+        return -1;
+    }
+    CloseHandle(token);
+
+    return elevation_info.TokenIsElevated;
+}
+
 /* check for native dll when running under wine */
 static BOOL is_native_dll( HMODULE module )
 {
@@ -222,8 +316,6 @@ static void print_version (void)
     static const char platform[] = "x86_64";
 #elif defined(__sparc__)
     static const char platform[] = "sparc";
-#elif defined(__ALPHA__)
-    static const char platform[] = "alpha";
 #elif defined(__powerpc__)
     static const char platform[] = "powerpc";
 #elif defined(__arm__)
@@ -232,11 +324,10 @@ static void print_version (void)
 # error CPU unknown
 #endif
     OSVERSIONINFOEX ver;
-    BOOL ext, wow64;
-    int is_win2k3_r2;
+    BOOL ext;
+    int is_win2k3_r2, is_admin, is_elevated;
     const char *(CDECL *wine_get_build_id)(void);
     void (CDECL *wine_get_host_version)( const char **sysname, const char **release );
-    BOOL (WINAPI *pIsWow64Process)(HANDLE hProcess, PBOOL Wow64Process);
     BOOL (WINAPI *pGetProductInfo)(DWORD, DWORD, DWORD, DWORD, DWORD *);
 
     ver.dwOSVersionInfoSize = sizeof(OSVERSIONINFOEX);
@@ -246,12 +337,18 @@ static void print_version (void)
 	if (!GetVersionEx ((OSVERSIONINFO *) &ver))
 	    report (R_FATAL, "Can't get OS version.");
     }
-    pIsWow64Process = (void *)GetProcAddress(GetModuleHandleA("kernel32.dll"),"IsWow64Process");
-    if (!pIsWow64Process || !pIsWow64Process( GetCurrentProcess(), &wow64 )) wow64 = FALSE;
-
-    xprintf ("    Platform=%s%s\n", platform, wow64 ? " (WOW64)" : "");
+    xprintf ("    Platform=%s%s\n", platform, is_wow64 ? " (WOW64)" : "");
     xprintf ("    bRunningUnderWine=%d\n", running_under_wine ());
     xprintf ("    bRunningOnVisibleDesktop=%d\n", running_on_visible_desktop ());
+    is_admin = running_as_admin ();
+    if (0 <= is_admin)
+    {
+        xprintf ("    Account=%s", is_admin ? "admin" : "non-admin");
+        is_elevated = running_elevated ();
+        if (0 <= is_elevated)
+            xprintf(", %s", is_elevated ? "elevated" : "not elevated");
+        xprintf ("\n");
+    }
     xprintf ("    Submitter=%s\n", email );
     if (description)
         xprintf ("    Description=%s\n", description );
@@ -612,6 +709,7 @@ run_test (struct wine_test* test, const char* subtest, HANDLE out_file, const ch
         status = run_ex (cmd, out_file, tempdir, 120000);
         heap_free (cmd);
         xprintf ("%s:%s done (%d)\n", test->name, subtest, status);
+        if (status) failures++;
     }
 }
 
@@ -1041,6 +1139,7 @@ usage (void)
 
 int main( int argc, char *argv[] )
 {
+    BOOL (WINAPI *pIsWow64Process)(HANDLE hProcess, PBOOL Wow64Process);
     char *logname = NULL, *outdir = NULL;
     const char *extract = NULL;
     const char *cp, *submit = NULL;
@@ -1050,6 +1149,9 @@ int main( int argc, char *argv[] )
     int i;
 
     if (!LoadStringA( 0, IDS_BUILD_ID, build_id, sizeof(build_id) )) build_id[0] = 0;
+
+    pIsWow64Process = (void *)GetProcAddress(GetModuleHandleA("kernel32.dll"),"IsWow64Process");
+    if (!pIsWow64Process || !pIsWow64Process( GetCurrentProcess(), &is_wow64 )) is_wow64 = FALSE;
 
     for (i = 1; i < argc && argv[i]; i++)
     {
@@ -1172,11 +1274,17 @@ int main( int argc, char *argv[] )
         if (!running_on_visible_desktop ())
             report (R_FATAL, "Tests must be run on a visible desktop");
 
-        if (!check_mount_mgr())
-            report (R_FATAL, "Mount manager not running, most likely your WINEPREFIX wasn't created correctly.");
+        if (running_under_wine())
+        {
+            if (!check_mount_mgr())
+                report (R_FATAL, "Mount manager not running, most likely your WINEPREFIX wasn't created correctly.");
 
-        if (!check_display_driver())
-            report (R_FATAL, "Unable to create a window, the display driver is not working.");
+            if (!check_wow64_registry())
+                report (R_FATAL, "WoW64 keys missing, most likely your WINEPREFIX wasn't created correctly.");
+
+            if (!check_display_driver())
+                report (R_FATAL, "Unable to create a window, the display driver is not working.");
+        }
 
         SetConsoleCtrlHandler(ctrl_handler, TRUE);
 
@@ -1213,7 +1321,13 @@ int main( int argc, char *argv[] )
                 DeleteFileA(logname);
                 exit (0);
             }
-            if (build_id[0] && nr_of_skips <= SKIP_LIMIT && !nr_native_dlls && !is_win9x &&
+            if (failures > FAILURES_LIMIT)
+                report( R_WARNING,
+                        "%d tests failed, there's probably something broken with your setup.\n"
+                        "You need to address this before submitting results.", failures );
+
+            if (build_id[0] && nr_of_skips <= SKIP_LIMIT && failures <= FAILURES_LIMIT &&
+                !nr_native_dlls && !is_win9x &&
                 report (R_ASK, MB_YESNO, "Do you want to submit the test results?") == IDYES)
                 if (!send_file (logname) && !DeleteFileA(logname))
                     report (R_WARNING, "Can't remove logfile: %u", GetLastError());
